requirements.txt
openfeature-sdk==0.7.1